// frontend/src/app/shared/models/recreacion-diversion.model.ts
export interface RecreacionDiversion {
  idEstab: string;
  subcategoriaRecreacion?: string;
  tipoServicio?: string;
  capacidadPersonas?: number;
  accesoAdaptado?: boolean;
  cantidadPiscinas?: number;
  acuaticoCubierto?: boolean;
  cantidadCancha?: number;
  tipoInstalacionesDeportivas?: string;
  cantidadSalasJuegos?: number;
  tieneSalaEscape?: boolean;
  observaciones?: string;
}